pyproject.toml
[project]
name = "pipeline6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.1.7",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
]
